backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
email-validator
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
slowapi
